Extract the multilingual aware fields to a concern
